Fix BTDownload precision loss: change getSize() from double to long
Torrent sizes exceed double precision above 2GB. Change return type
from double to long across Transfer, BTDownload, and all callers
in desktop and android modules.

Also: PerformersHelper now filters hidden files alongside pad files,
and CreateTorrentDialog TODO is converted to upstream note.

+++ android/src/main/java/com/frostwire/android/gui/dialogs/NewTransferDialog.java
@@ -149,7 +149,7 @@ public class NewTransferDialog extends AbstractDialog {
     private static final class SearchResultData implements Serializable {
 
         private final String displayName;
-        private final double size;
+        private final long size;
 
         public SearchResultData(FileSearchResult sr) {
             this.displayName = sr.getDisplayName();
@@ -160,7 +160,7 @@ public class NewTransferDialog extends AbstractDialog {
             return displayName;
         }
 
-        public double getSize() {
+        public long getSize() {
             return size;
         }
     }
