+++ common/src/main/java/com/frostwire/transfers/Transfer.java
@@ -35,7 +35,7 @@ public interface Transfer {
 
     File previewFile();
 
-    double getSize();
+    long getSize();
 
     Date getCreated();
